+++ packages/pangea-multipass/pangea_multipass/__init__.py
@@ -16,7 +16,10 @@ from .core import (
     generate_id,
     get_document_metadata,
 )
+from .dropbox_reader import DropboxReader
 from .github_reader import GitHubReader
 from .gitlab_reader import GitLabReader
+from .oauth import OauthFlow
 from .slack_reader import SlackReader
 from .sources import *
+from .utils import *