+++ packages/pangea-multipass-llama-index/CHANGELOG.md
@@ -7,9 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-### Added 
+### Added
 
 - GitLabReader and GitLabProcessor
+- Dropbox processor
 
 ### Fixed